Add Suspense handling to history page
